### 2022-06-16 #4482 @arfon
@FlukeAndFeather – it looks like the ORCID for one of the authors is incorrect: https://github.com/openjournals/joss-papers/runs/6922434417?check_suite_focus=true#step:3:95

Also, given the size of this submission, I will need to do a scope review to verify with the wider editorial team that this submission meets our [substantial scholarly effort criterion](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). This will likely take a week or so.
