+++ vignettes/references.bib
@@ -92,7 +92,7 @@
 	pages = {1639--1651},
 	volume = {11},
 	number = {12},
-	doi = {https://doi.org/10.1111/2041-210X.13491},
+	doi = {10.1111/2041-210X.13491},
 	url = {https://besjournals.onlinelibrary.wiley.com/doi/abs/10.1111/2041-210X.13491},
 	note = {{\_}eprint: https://besjournals.onlinelibrary.wiley.com/doi/pdf/10.1111/2041-210X.13491},
 	langid = {en}
