### 2022-06-16 #4482 @FlukeAndFeather
> @FlukeAndFeather – it looks like the ORCID for one of the authors is incorrect: https://github.com/openjournals/joss-papers/runs/6922434417?check_suite_focus=true#step:3:95

Sorry, not sure how that happened. I corrected both authors' ORCIDs [550102f](https://github.com/FlukeAndFeather/rstickleback/commit/550102fb6a8bf868757f94bd9c9e77d285f2d3c9).

> Also, given the size of this submission, I will need to do a scope review to verify with the wider editorial team that this submission meets our [substantial scholarly effort criterion](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). This will likely take a week or so.

Understood. There might have been a better way for me to submit this software, as it also includes the [`stickleback`](https://github.com/FlukeAndFeather/stickleback) Python package, which is a separate repository. I decided to submit the `rstickleback` repo because that's the software package intended for end-users. If it helps in your decision making, I presented this software at the 7th International Bio-logging Symposium in October 2021 and it's the basis for an upcoming manuscript we're preparing for *Methods in Ecology and Evolution*.

This is my first time submitting to JOSS, so apologies in advance for any headaches I cause!
