### 2022-07-15 #4482 @chartgerink
Thanks @ianjonsen for letting us know! 😊

@chfleming and @rociojoo - might one of you be interested to review this paper on bio-logging using machine learning?

Find [the paper on GitHub](https://github.com/openjournals/joss-papers/blob/joss.04482/joss.04482/10.21105.joss.04482.pdf), and [the source code](https://github.com/FlukeAndFeather/rstickleback) as well.
