### 2022-07-18 #4482 @chartgerink
Thanks @chfleming! Once review commences, a new issue will be added where you can start [your review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html). You then go through the items and you may leave issues on the original repository.

I'm always around to answer questions so feel free to ping me! For now, you don't need to do anything until I find a second reviewer :)
