### 2022-07-22 #4482 @chartgerink
Dear @pboesu @robohyun66 @skummerf - might one/some of you be interested to review this paper on bio-logging using machine learning?

Find [the paper on GitHub](https://github.com/openjournals/joss-papers/blob/joss.04482/joss.04482/10.21105.joss.04482.pdf), and [the source code](https://github.com/FlukeAndFeather/rstickleback) as well.

I would greatly appreciate any help you could provide in reviewing for JOSS. If you're unavailable, thank you for reading and feel free to let me know as well 😊
