### 2022-08-09 #4662 @editorialbot
[REVIEW]: rstickleback: supervised behavior detection in bio-logging data
**Submitting author:** @FlukeAndFeather (<a href="http://orcid.org/0000-0002-6302-905X">Max Czapanskiy</a>)
**Repository:** https://github.com/FlukeAndFeather/rstickleback
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.3.0
