### 2022-08-29 #4662 @jgalgarra
Hi,

I got stuck installing the Python package stickleback. I followed the installation instructions, launched the devtools::install_github("FlukeAndFeather/rstickleback") command from RStudio and everything run smoothly. Then, RStudio requested permission to install miniconda, and stickleback installation failed but it was no big deal since the instructions specified that is available with pip. The problem is that "pip install stickleback" is raising an exception. I have tried to install it in my Python 3.8.8 environments, both from Windows and Linux and I always get the same error log (attached). I would like to know @chfleming if you have had the same problem, and @chartgerink if I could contact the authors to request their support. Rstickleback is a wrapper of stickleback, and so they are not responsible of this particular installation problem, but
without stickleback is impossible to move on the review.

[exceptionlog.txt](https://github.com/openjournals/joss-reviews/files/9445137/exceptionlog.txt)
