### 2022-08-30 #4662 @FlukeAndFeather
I see the issues in the original repo. Thanks @jgalgarra for the thorough review! I appreciate your points about making the tool as easy to use as possible for non-Python users. That is a core goal for the software and it helps to have someone else's perspective on likely challenges. I'll get started on those issues later this week.
