### 2022-09-05 FlukeAndFeather/rstickleback#38 @tretherington
JOSS review: installation instructions
For some context I'm not a trained computer scientist so I do sometimes struggle with more complicated computer stuff, but I have used R and Python in ecological settings for many years now.  I say that as I think I am more bilingual than the average ecologist so am more likely to be able to problem solve, so if I get stuck I think that is a reasonable indicator that others may get stuck.  I think it's also worth highlighting that when I try to use Python in my ecology work I do get pushback and almost fear from colleagues, so with a view to making your package as useful as possible, I really think to need to make the installation as painless as possible otherwise you're probably going to lose possible users very quickly.

OK, that said, I have failed to install rstickleback, and I'm unsure how to proceed.

So I did:
```
devtools::install_github("FlukeAndFeather/rstickleback")
```
as per the README installation instructions (I had to repeat a couple of times to get all the dependencies to install, but that might have been a problem at my end).  I then tried to load the package and was prompted to install Miniconda, but I already have that installed so I opted for no, so the install aborted.

```
> library(rstickleback)
No non-system installation of Python could be found.
Would you like to download and install Miniconda?
Miniconda is an open source environment management system for Python.
See https://docs.conda.io/en/latest/miniconda.html for more details.

Would you like to install Miniconda? [Y/n]: n
Installation aborted.
Python package stickleback not found.
Warning message:
S3 methods ‘base::as.data.frame.Events’, ‘base::as.data.frame.Outcomes’, ‘base::as.data.frame.Predictions’, ‘base::as.data.frame.Sensors’ were declared in NAMESPACE but not found
```

Then I thought perhaps I am installation a version of Miniconda that some how sits 'inside' R and would be seperate to my existing Miniconda install?  This is the kind of subtlety that I think would be worth explaining to someone like me who has not called Python from R before.  So I thought I'd try and load stickleback to prompt the Miniconda install, but this time no prompt came up.

```
> library(rstickleback)
>
>
```
A reload of RStudio to try again and I'm still not getting a Miniconda install prompt:
```
> library(rstickleback)
Python package stickleback not found.
Warning message:
S3 methods ‘base::as.data.frame.Events’, ‘base::as.data.frame.Outcomes’, ‘base::as.data.frame.Predictions’, ‘base::as.data.frame.Sensors’ were declared in NAMESPACE but not found
```
So I am now at a loss as to how to proceed as the Installation instructions in the README just say to do ```devtools::install_github("FlukeAndFeather/rstickleback")``` and nothing else.

I completely acknowledge that I may have done something 'silly' here, but in the hope of making your package as fool proof as possibly, this fool has certainly got stuck.
