### 2022-09-05 #4662 @tretherington
Apologies @FlukeAndFeather for the slow start on this.  I've had a first go but have got stuck on installation - something that I think will need to be well documented to get an R using ecologist using your package.  I've raised some issues in the hope they help you to improve things. 😄
