### 2022-09-05 FlukeAndFeather/rstickleback#38 @chfleming
For me, miniconda installed, and the python packages downloaded, but then I got an error quickly into the installation of python packages:

```
Attempting uninstall: setuptools
    Found existing installation: setuptools 65.3.0
    Uninstalling setuptools-65.3.0:
      Successfully uninstalled setuptools-65.3.0
  Attempting uninstall: numpy
    Found existing installation: numpy 1.23.2
    Uninstalling numpy-1.23.2:
      Successfully uninstalled numpy-1.23.2
ERROR: Could not install packages due to an OSError: [WinError 5] Access is denied: 'C:\\Users\\Christen\\AppData\\Local\\r-miniconda\\envs\\r-reticulate\\Lib\\site-packages\\~umpy\\core\\_multiarray_tests.cp38-win_amd64.pyd'
Consider using the `--user` option or check the permissions.
```

I assume this is because I did not install as Administrator in Windows? I installed as user "Christen" with permissions to the listed pyd file, so the error is somewhat confusing.
