### 2022-09-05 FlukeAndFeather/rstickleback#38 @chfleming
Never mind, I restarted R and tried again and it installed. It must have been a random Windows file-lock error.
