### 2022-09-05 FlukeAndFeather/rstickleback#39 @chfleming
JOSS review: quick start
I'm stuck on the first line after loading the library. Does there need to be another package in the dependencies?

```
> library(rstickleback)
> c(lunge_sensors, lunge_events) %<-% load_lunges()
Error in `.rowNamesDF<-`(x, value = value) :
  duplicate 'row.names' are not allowed
In addition: Warning message:
non-unique values when setting 'row.names': ‘2018-09-05 11:55:52’, ‘2018-09-05 11:55:53’, ‘2018-09-05 11:55:54’, ‘2018-09-05 11:55:55’, ‘2018-09-05 11:55:56’, ‘2018-09-05 11:55:57’, ‘2018-09-05 11:55:58’, ‘2018-09-05 11:55:59’, ‘2018-09-05 11:56:00’, ‘2018-09-05 11:56:01’, ‘2018-09-05 11:56:02’, ‘2018-09-05 11:56:03’, ‘2018-09-05 11:56:04’, ‘2018-09-05 11:56:05’, ‘2018-09-05 11:56:06’, ‘2018-09-05 11:56:07’, ‘2018-09-05 11:56:08’, ‘2018-09-05 11:56:09’, ‘2018-09-05 11:56:10’, ‘2018-09-05 11:56:11’, ‘2018-09-05 11:56:12’, ‘2018-09-05 11:56:13’, ‘2018-09-05 11:56:14’, ‘2018-09-05 11:56:15’, ‘2018-09-05 11:56:16’, ‘2018-09-05 11:56:17’, ‘2018-09-05 11:56:18’, ‘2018-09-05 11:56:19’, ‘2018-09-05 11:56:20’, ‘2018-09-05 11:56:21’, ‘2018-09-05 11:56:22’, ‘2018-09-05 11:56:23’, ‘2018-09-05 11:56:24’, ‘2018-09-05 11:56:25’, ‘2018-09-05 11:56:26’, ‘2018-09- [... truncated]
```
