### 2022-09-06 #4662 @chfleming
I was able to get through the installation, but then I got an error on the first line of the quick start example after loading the library: https://github.com/FlukeAndFeather/rstickleback/issues/39
