### 2022-09-06 #4662 @tretherington
Still stuck on installation: [https://github.com/FlukeAndFeather/rstickleback/issues/39](https://github.com/FlukeAndFeather/rstickleback/issues/38)
