### 2022-09-06 FlukeAndFeather/rstickleback#38 @FlukeAndFeather
Thank you both for finding these installation errors. I think "Windows with pre-installed Miniconda" is a system configuration I didn't test, and there seem to be some other unrelated issues. I'm currently on vacation, but @A-JM and I have plans to identify and fix the errors later this month. If you'd like to pause your review, I'd be happy to alert you when we've made the necessary changes. Of course feel free to continue if you solve the installation issues on your own - I just don't want to waste your time. Thanks!
