### 2022-09-06 #4662 @FlukeAndFeather
[Dropping a link](https://github.com/FlukeAndFeather/rstickleback/issues/38#issuecomment-1238633970) to my reply on the rstickleback issue. In short, the installation issues seem to be a blocking problem. I suggested @chfleming and @tretherington pause their review while @A-JM and I fix these critical issues. I am currently on vacation but will work on the fix later this month. Thanks everyone!
