### 2022-10-11 #4482 @junbinzhao
Sorry I am on vacation now and can't find time. But good luck with the process. JunbinSent from my Galaxy
-------- Original message --------From: Chris Hartgerink ***@***.***> Date: 29/07/2022  10:55  (GMT+01:00) To: openjournals/joss-reviews ***@***.***> Cc: Junbin Zhao ***@***.***>, Mention ***@***.***> Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: rstickleback: supervised behavior detection in bio-logging data (Issue #4482)

Dear @SteveViss @junbinzhao @tretherington @jgalgarra - might one/some of you be interested to review this paper on bio-logging using machine learning?

Find the paper on GitHub, and the source code as well.

I would greatly appreciate any help you could provide in reviewing for JOSS. If you're unavailable, thank you for reading and feel free to let me know as well 😊
