### 2023-05-05 #4662 @FlukeAndFeather
Hello @kthyng. Yes, I'm afraid I have been swamped with a new position and I apologize for the lack of movement here. I think the amount of troubleshooting required for seamless Python installation via R is too much for me right now. Let's go with option 1 and withdraw the paper for now. My thanks to the reviewers for their feedback and I'll hopefully resubmit in the next couple months.
